Major Refactor Milestone #1 - filename and include changes
