+++ gui/about_dialog.h
