+++ gui/nb_selector.ui
