+++ gui/remove_plugin_dialog.h
