+++ gui/remove_plugin_dialog.ui
