+++ src/image_loader.h
