+++ src/light_source.h
