+++ styles.qrc
@@ -1,8 +1,8 @@
 <RCC>
     <qresource prefix="/">
-        <file>styles/spinboxslider.qss</file>
-        <file>styles/spinboxslider2.qss</file>
+        <file>styles/spin_box_slider.qss</file>
+        <file>styles/spin_box_slider2.qss</file>
         <file>styles/classic.qss</file>
-        <file>styles/dark-blue.qss</file>
+        <file>styles/dark_blue.qss</file>
     </qresource>
 </RCC>
