+++ styles/dark_blue.qss
