+++ styles/spin_box_slider.qss
