+++ styles/spin_box_slider2.qss
